before action inherited class created
