+++ lib/hackney/pdf/preview_generator.rb
@@ -19,8 +19,8 @@ module Hackney
         @logo = File.open(LOGO_PATH).read
       end
 
-      def execute(letter_params:)
-        @letter = Hackney::ServiceCharge::Letter.new(letter_params)
+      def execute(letter_params:, template_id:)
+        @letter = Hackney::ServiceCharge::Letter.build_letter(letter_params: letter_params, template_id: template_id)
 
         @sender_address = ERB.new(File.open(SENDER_ADDRESS_PATH).read).result(binding)
